Refile filter tests by target module
braincell/filter/ had five source modules and no test file named after any
of them. Refile by what each test actually exercises:

  filter_branch_filters_test.py -> region_test.py (BranchInFilter and
      BranchRangeFilter are both defined in region.py)
  filter_region_test.py         -> region_test.py, with FilterModuleAllTest
      split by method across region_test.py, locset_test.py, and a new
      helper_test.py
  filter_vis_test.py            -> morph/morphology_test.py (10 tests for
      Morphology.select/.vis2d/.vis3d), morph/branch_test.py (1 test for the
      Branch-view rejection contract), and vis/plot3d_test.py (3 tests for
      a filter region flowing into plot3d)

filter_vis_test.py never belonged in filter/: its subject is
morph/morphology.py, and braincell.filter symbols are only inputs.

The two source files each defined a module-level _build_tree() with
different geometry (3 branches vs 4). Both land in region_test.py, so they
are renamed _soma_dend_axon_tree and _soma_dend_axon_tuft_tree rather than
hoisted; no filter/_testing.py is needed after all, since every remaining
helper is used by exactly one module.

Separately, vis/_test_helper.py held a single class, FakeBackend, which
AGENTS.md says belongs in the package's _testing.py. Folded in and deleted;
its three importers now read from braincell.vis._testing.

Collected test count is unchanged at 2276.
